serial: stm32: Selectively update regs in set_termios
When set_termios is called, USART is stopped for a while
(by UE=0) and all characters currently in FIFO are lost.
It can be seen when busybox shell initialy starts on
/dev/console for example.
Detect which bits are about to change and compare to
masks (obtained from datasheets) to decide whether
we need to stop USART. We can change other bits
on the fly (RTOR, interrupt masks..).

It handles mainly two cases:
- USART handoff from bootloader to kernel where RTIE
  and RE bits are typically set
- ioctl changes which changes only TTY values and
  serial link registers are unmodified

Tested on STM32MP1.

Signed-off-by: Martin Devera <devik@eaxlabs.cz>